requirements.txt
requests>=2.32.0
pydantic>=2.8.0
python-dotenv>=1.0.1
redis>=5.0.7
tenacity>=9.0.0
